// src/lib/components/_SourceTable/interface.ts
export type IRowActions = Array<React.ReactElement<IRowActionProps>>;

export type ISourceTableBaseRowActions = IRowActions | ((event: TableRowEventData) => IRowActions);
// ...
export interface TableRowEventData {
  rowId: string;
  rowData: any;
  rowIndex: number;
  originEvent?: Event;
  params: ActionCellRendererClassParams;
}